@@ -6,9 +6,9 @@
     "templates/*"
   ],
   "devDependencies": {
-    "eslint": "^8.35.0",
-    "eslint-config-lxsmnsyc": "^0.5.0",
-    "lerna": "^6.4.1",
-    "typescript": "^4.9.5"
+    "eslint": "^8.51.0",
+    "eslint-config-lxsmnsyc": "^0.6.5",
+    "lerna": "^7.3.0",
+    "typescript": "^5.2.2"
   }
 }
