@@ -6,20 +6,20 @@
     "src"
   ],
   "engines": {
-    "node": ">=12"
+    "node": ">=16"
   },
   "license": "MIT",
   "keywords": [
     "pridepack"
   ],
   "devDependencies": {
-    "@types/node": "^18.14.1",
-    "eslint": "^8.35.0",
-    "eslint-config-lxsmnsyc": "^0.5.0",
+    "@types/node": "^20.8.3",
+    "eslint": "^8.51.0",
+    "eslint-config-lxsmnsyc": "^0.6.5",
     "pridepack": "2.4.4",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
-    "vitest": "^0.28.1"
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vitest": "^0.34.6"
   },
   "scripts": {
     "test": "vitest"
