@@ -6,24 +6,24 @@
     "src"
   ],
   "engines": {
-    "node": ">=12"
+    "node": ">=16"
   },
   "license": "MIT",
   "keywords": [
     "pridepack"
   ],
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/preact": "^3.2.3",
-    "@types/node": "^18.14.1",
-    "eslint": "^8.35.0",
-    "eslint-config-lxsmnsyc": "^0.5.0",
-    "jsdom": "^21.1.0",
-    "preact": "^10.11.3",
+    "@types/node": "^20.8.3",
+    "eslint": "^8.51.0",
+    "eslint-config-lxsmnsyc": "^0.6.5",
+    "jsdom": "^22.1.0",
+    "preact": "^10.17.1",
     "pridepack": "2.4.4",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
-    "vitest": "^0.28.1"
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vitest": "^0.34.6"
   },
   "peerDependencies": {
     "preact": "^10.0"
