@@ -6,27 +6,27 @@
     "src"
   ],
   "engines": {
-    "node": ">=12"
+    "node": ">=16"
   },
   "license": "MIT",
   "keywords": [
     "pridepack"
   ],
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@types/node": "^18.14.1",
-    "@types/react": "^18.0.27",
-    "eslint": "^8.35.0",
-    "eslint-config-lxsmnsyc": "^0.5.0",
-    "jsdom": "^21.1.0",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
+    "@types/node": "^20.8.3",
+    "@types/react": "^18.2.25",
+    "eslint": "^8.51.0",
+    "eslint-config-lxsmnsyc": "^0.6.5",
+    "jsdom": "^22.1.0",
     "pridepack": "2.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-test-renderer": "^18.2.0",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
-    "vitest": "^0.28.1"
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vitest": "^0.34.6"
   },
   "peerDependencies": {
     "react": "^16.8 || ^17.0 || ^18.0",
