@@ -8,10 +8,8 @@ module.exports = {
     "tsconfigRootDir": __dirname,
   },
   "rules": {
-    "no-param-reassign": "off",
     "@typescript-eslint/no-unsafe-assignment": "off",
     "@typescript-eslint/no-unsafe-argument": "off",
-    "no-await-in-loop": "off",
-    "no-restricted-syntax": "off"
+    "no-await-in-loop": "off"
   },
 };
