@@ -1,6 +1,6 @@
-import { Abortable } from 'events';
-import { Mode, ObjectEncodingOptions, OpenMode } from 'fs';
-import { Stream } from 'stream';
+import type { Abortable } from 'events';
+import type { Mode, ObjectEncodingOptions, OpenMode } from 'fs';
+import type { Stream } from 'stream';
 import path from 'path';
 import fs from 'fs/promises';
 
@@ -38,7 +38,7 @@ export function checkPath(filePath: string): void {
   }
 }
 
-export async function makeDir(dir: string, mode = 0o777) {
+export async function makeDir(dir: string, mode = 0o777): Promise<string | undefined> {
   checkPath(dir);
   return fs.mkdir(dir, {
     mode,
@@ -46,7 +46,7 @@ export async function makeDir(dir: string, mode = 0o777) {
   });
 }
 
-export async function pathExists(filePath: string) {
+export async function pathExists(filePath: string): Promise<boolean> {
   try {
     await fs.access(filePath);
     return true;
@@ -70,17 +70,17 @@ export async function outputFile(
         } & Abortable)
       | BufferEncoding
       | null,
-) {
+): Promise<void> {
   const dir = path.dirname(file);
   if (!await pathExists(dir)) {
     await makeDir(dir);
   }
-  return fs.writeFile(file, data, encoding);
+  await fs.writeFile(file, data, encoding);
 }
 
 function stringify(
-  obj: any,
-) {
+  obj: unknown,
+): string {
   const str = JSON.stringify(obj, undefined, 2);
 
   return `${str.replace(/\n/g, '\n')}\n`;
@@ -88,9 +88,9 @@ function stringify(
 
 export async function outputJson(
   file: string,
-  data: any,
-) {
-  return outputFile(file, stringify(data));
+  data: unknown,
+): Promise<void> {
+  await outputFile(file, stringify(data));
 }
 
 export async function readJson<T>(
