@@ -1,23 +1,24 @@
-import { BuildResult, BuildContext } from 'esbuild';
+import type { BuildResult, BuildContext } from 'esbuild';
 import buildBundle from '../core/build-bundle';
-import { PridepackConfig } from '../core/default-config';
+import type { PridepackConfig } from '../core/default-config';
 import runESBuild from './run-esbuild';
+import type { Task } from './run-task';
 
 interface BuildCall {
   (incremental: false): Promise<BuildResult>;
   (incremental: true): Promise<BuildContext>;
 }
 
-export default function runBuild(
+export default async function runBuild(
   config: PridepackConfig,
   incremental: boolean,
   isDev: boolean,
   isESM: boolean,
-) {
+): Promise<Task<void>> {
   const format = isESM ? 'ESM' : 'CommonJS';
   const mode = isDev ? 'Development' : 'Production';
   return runESBuild(
-    ((inc) => buildBundle(config, inc, isDev, isESM)) as BuildCall,
+    (async (inc) => buildBundle(config, inc, isDev, isESM)) as BuildCall,
     incremental,
     {
       pending: `Building ${format} ${mode} output...`,
