@@ -7,7 +7,7 @@ import runTask from './run-task';
 
 export default async function runInitCommand(): Promise<void> {
   const templateName = await chooseTemplate();
-  const task = await runTask(() => copyFromTemplate(templateName.template, '.'), {
+  const task = await runTask(async () => copyFromTemplate(templateName.template, '.'), {
     pending: `Copying from template '${templateName.template}'...`,
     success: `Copied from template '${templateName.template}'!`,
     failure: `Failed to copy from template '${templateName.template}'.`,
